refactor(http): extract error telemetry capture into a private module
